Most of comments done.
